# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Industrielle RH API",
    "description": "API pour la plateforme de recrutement Industrielle RH - Gestion des candidatures, recrutements, calculs salariaux et plus",
    "version": "1.0.0",
    "contact": {
      "name": "Industrielle RH",
      "email": "info@industriellerh.ca",
      "url": "https://industriellerh.ca"
    },
    "license": {
      "name": "Propriétaire",
      "url": "https://industriellerh.ca/terms"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Serveur de développement"
    },
    {
      "url": "https://industriellerh.ca/api",
      "description": "Serveur de production"
    }
  ],
  "paths": {
    "/contacts": {
      "get": {
        "tags": ["Contacts"],
        "summary": "Récupérer tous les contacts",
        "description": "Récupère la liste paginée des contacts avec possibilité de recherche et filtrage",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Numéro de page",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Nombre d'éléments par page",
            "schema": {
              "type": "integer",
              "default": 10
            }
          },
          {
            "name": "search",
            "in": "query",
            "description": "Terme de recherche",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "Filtrer par statut",
            "schema": {
              "type": "string",
              "enum": ["nouveau", "en_cours", "traite", "ferme"]
            }
          },
          {
            "name": "priority",
            "in": "query",
            "description": "Filtrer par priorité",
            "schema": {
              "type": "string",
              "enum": ["basse", "moyenne", "haute", "urgente"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des contacts récupérée avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Contact"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/PaginationMeta"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Contacts"],
        "summary": "Créer un nouveau contact",
        "description": "Crée un nouveau contact dans le système",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContactCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Contact créé avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Contact"
                }
              }
            }
          },
          "400": {
            "description": "Données invalides",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/salary-calculations": {
      "get": {
        "tags": ["Calculs Salariaux"],
        "summary": "Récupérer tous les calculs salariaux",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 10
            }
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "industry",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "location",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des calculs salariaux",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/SalaryCalculation"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/PaginationMeta"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Calculs Salariaux"],
        "summary": "Créer un nouveau calcul salarial",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SalaryCalculationCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Calcul salarial créé avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/SalaryCalculation"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Données invalides",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string"
                    },
                    "errors": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/tax-calculations": {
      "get": {
        "tags": ["Calculs d'Impôts"],
        "summary": "Récupérer tous les calculs d'impôts",
        "responses": {
          "200": {
            "description": "Liste des calculs d'impôts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/TaxCalculation"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Calculs d'Impôts"],
        "summary": "Créer un nouveau calcul d'impôt",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaxCalculationCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Calcul d'impôt créé avec succès"
          }
        }
      }
    },
    "/mortgage-calculations": {
      "get": {
        "tags": ["Calculs Hypothécaires"],
        "summary": "Récupérer tous les calculs hypothécaires",
        "responses": {
          "200": {
            "description": "Liste des calculs hypothécaires"
          }
        }
      },
      "post": {
        "tags": ["Calculs Hypothécaires"],
        "summary": "Créer un nouveau calcul hypothécaire",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MortgageCalculationCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Calcul hypothécaire créé avec succès"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Contact": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Identifiant unique du contact"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Date de création"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Date de dernière modification"
          },
          "firstName": {
            "type": "string",
            "description": "Prénom",
            "example": "Jean"
          },
          "lastName": {
            "type": "string",
            "description": "Nom de famille",
            "example": "Dupont"
          },
          "companyName": {
            "type": "string",
            "nullable": true,
            "description": "Nom de l'entreprise",
            "example": "Acme Corp"
          },
          "jobTitle": {
            "type": "string",
            "nullable": true,
            "description": "Titre du poste",
            "example": "Directeur RH"
          },
          "workEmail": {
            "type": "string",
            "format": "email",
            "description": "Adresse email professionnelle",
            "example": "jean.dupont@acme.com"
          },
          "workPhone": {
            "type": "string",
            "description": "Téléphone professionnel",
            "example": "819-123-4567"
          },
          "postalCode": {
            "type": "string",
            "nullable": true,
            "description": "Code postal",
            "example": "J1A 1A1"
          },
          "message": {
            "type": "string",
            "description": "Message du contact",
            "example": "Nous recherchons un ingénieur en mécanique..."
          },
          "status": {
            "type": "string",
            "enum": ["nouveau", "en_cours", "traite", "ferme"],
            "description": "Statut du contact",
            "example": "nouveau"
          },
          "priority": {
            "type": "string",
            "enum": ["basse", "moyenne", "haute", "urgente"],
            "description": "Priorité du contact",
            "example": "moyenne"
          }
        },
        "required": ["id", "firstName", "lastName", "workEmail", "workPhone", "message", "status", "priority"]
      },
      "ContactCreate": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100,
            "example": "Jean"
          },
          "lastName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100,
            "example": "Dupont"
          },
          "companyName": {
            "type": "string",
            "maxLength": 200,
            "example": "Acme Corp"
          },
          "jobTitle": {
            "type": "string",
            "maxLength": 150,
            "example": "Directeur RH"
          },
          "workEmail": {
            "type": "string",
            "format": "email",
            "maxLength": 320,
            "example": "jean.dupont@acme.com"
          },
          "workPhone": {
            "type": "string",
            "maxLength": 20,
            "example": "819-123-4567"
          },
          "postalCode": {
            "type": "string",
            "maxLength": 10,
            "example": "J1A 1A1"
          },
          "message": {
            "type": "string",
            "minLength": 1,
            "maxLength": 2000,
            "example": "Nous recherchons un ingénieur en mécanique..."
          },
          "status": {
            "type": "string",
            "enum": ["nouveau", "en_cours", "traite", "ferme"],
            "default": "nouveau"
          },
          "priority": {
            "type": "string",
            "enum": ["basse", "moyenne", "haute", "urgente"],
            "default": "moyenne"
          }
        },
        "required": ["firstName", "lastName", "workEmail", "workPhone", "message"]
      },
...
      "SalaryCalculation": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "jobTitle": {
            "type": "string",
            "example": "Ingénieur Logiciel"
          },
          "industry": {
            "type": "string",
            "example": "Technologie"
          },
          "location": {
            "type": "string",
            "example": "Montréal, QC"
          },
          "experience": {
            "type": "integer",
            "minimum": 0,
            "example": 5
          },
          "education": {
            "type": "string",
            "nullable": true,
            "example": "Baccalauréat"
          },
          "skills": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": ["JavaScript", "React", "Node.js"]
          },
          "company": {
            "type": "string",
            "nullable": true,
            "example": "Tech Corp"
          },
          "salaryRange": {
            "type": "string",
            "example": "70000-85000"
          },
          "email": {
            "type": "string",
            "format": "email",
            "nullable": true,
            "example": "user@example.com"
          }
        },
        "required": ["jobTitle", "industry", "location", "experience", "salaryRange"]
      },
      "SalaryCalculationCreate": {
        "type": "object",
        "properties": {
          "jobTitle": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200,
            "description": "Titre du poste"
          },
          "industry": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200,
            "description": "Secteur d'activité"
          },
          "location": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200,
            "description": "Localisation"
          },
          "experience": {
            "type": "integer",
            "minimum": 0,
            "maximum": 50,
            "description": "Années d'expérience"
          },
          "education": {
            "type": "string",
            "maxLength": 100,
            "description": "Niveau d'éducation"
          },
          "skills": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Compétences"
          },
          "company": {
            "type": "string",
            "maxLength": 200,
            "description": "Entreprise"
          },
          "salaryRange": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100,
            "description": "Fourchette salariale"
          },
          "email": {
            "type": "string",
            "format": "email",
            "maxLength": 320,
            "description": "Email"
          }
        },
        "required": ["jobTitle", "industry", "location", "experience", "salaryRange"]
      },
      "TaxCalculation": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "TaxCalculationCreate": {
        "type": "object",
        "properties": {
          "annualSalary": {
            "type": "number",
            "minimum": 0,
            "description": "Salaire annuel"
          },
          "province": {
            "type": "string",
            "description": "Province"
          },
          "deductions": {
            "type": "number",
            "minimum": 0,
            "description": "Déductions"
          }
        },
        "required": ["annualSalary", "province"]
      },
      "MortgageCalculationCreate": {
        "type": "object",
        "properties": {
          "propertyValue": {
            "type": "number",
            "minimum": 0,
            "description": "Valeur de la propriété"
          },
          "downPayment": {
            "type": "number",
            "minimum": 0,
            "description": "Mise de fonds"
          },
          "interestRate": {
            "type": "number",
            "minimum": 0,
            "description": "Taux d'intérêt"
          },
          "amortizationPeriod": {
            "type": "integer",
            "minimum": 1,
            "description": "Période d'amortissement en années"
          }
        },
        "required": ["propertyValue", "downPayment", "interestRate", "amortizationPeriod"]
      },
...
      "PaginationMeta": {
        "type": "object",
        "properties": {
          "total": {
            "type": "integer",
            "description": "Nombre total d'éléments"
          },
          "page": {
            "type": "integer",
            "description": "Page actuelle"
          },
          "limit": {
            "type": "integer",
            "description": "Nombre d'éléments par page"
          },
          "totalPages": {
            "type": "integer",
            "description": "Nombre total de pages"
          }
        },
        "required": ["total", "page", "limit", "totalPages"]
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Message d'erreur"
          },
          "message": {
            "type": "string",
            "description": "Description détaillée de l'erreur"
          },
          "statusCode": {
            "type": "integer",
            "description": "Code de statut HTTP"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}